requirements.txt
python-telegram-bot==13.7
pymongo==3.12.0
mercadopago==1.12.0
Flask==2.3.2
gunicorn==20.1.0
schedule==1.1.0
